move Web APIs to cli/js/web/
