+++ cli/js/web/form_data.ts
@@ -2,8 +2,8 @@
 import * as domTypes from "./dom_types.ts";
 import * as blob from "./blob.ts";
 import * as domFile from "./dom_file.ts";
-import { DomIterableMixin } from "./mixins/dom_iterable.ts";
-import { requiredArguments } from "./util.ts";
+import { DomIterableMixin } from "./dom_iterable.ts";
+import { requiredArguments } from "../util.ts";
 
 const dataSymbol = Symbol("data");
