+++ cli/js/web/headers.ts
@@ -1,8 +1,8 @@
 // Copyright 2018-2020 the Deno authors. All rights reserved. MIT license.
 import * as domTypes from "./dom_types.ts";
-import { DomIterableMixin } from "./mixins/dom_iterable.ts";
-import { requiredArguments } from "./util.ts";
-import { customInspect } from "./console.ts";
+import { DomIterableMixin } from "./dom_iterable.ts";
+import { requiredArguments } from "../util.ts";
+import { customInspect } from "../console.ts";
 
 // From node-fetch
 // Copyright (c) 2016 David Frank. MIT License.
