+++ cli/js/web/location.ts
@@ -1,6 +1,6 @@
 // Copyright 2018-2020 the Deno authors. All rights reserved. MIT license.
 import { URL } from "./url.ts";
-import { notImplemented } from "./util.ts";
+import { notImplemented } from "../util.ts";
 import { Location } from "./dom_types.ts";
 
 export class LocationImpl implements Location {
