+++ cli/js/web/streams/shared-internals.ts
